+++ src/main/cpp/querytable/query_field.cpp
@@ -19,7 +19,7 @@ QueryField::QueryField(emp::Integer val, int length, int field_num)
       value_(types::TypeId::ENCRYPTED_INTEGER64, val, length) {}
 
 QueryField::QueryField(double val, int field_num)
-    : field_num_(field_num), value_(types::TypeId::FLOAT64, val) {}
+    : field_num_(field_num), value_(types::TypeId::FLOAT32, val) {}
 
 types::Value *QueryField::GetMutableValue() { return &value_; }
